split constants into different files
