+++ src/incentives/bin/server.js
@@ -14,7 +14,7 @@ const helmet = require("helmet");
 const isDev = process.env.NODE_ENV === "development";
 const isProd = process.env.NODE_ENV === "production";
 const options = { mongooseConnection: mongoose.connection };
-require("@bin/cronJob");
+require("@bin/refresh-sims-job");
 const log4js = require("log4js");
 const debug = require("debug")("auth-service:server");
 const isEmpty = require("is-empty");
