+++ src/incentives/config/constants.js
@@ -1,145 +1,8 @@
-const devConfig = {
-  MONGO_URI: process.env.DEV_MONGO_URI,
-  DB_NAME: process.env.DEV_MONGO_DB,
-  ENVIRONMENT: "DEVELOPMENT ENVIRONMENT",
-  REDIS_SERVER: process.env.DEV_REDIS_SERVER,
-  REDIS_PORT: process.env.DEV_REDIS_PORT,
-  XENTE_BASE_URL: process.env.DEV_XENTE_BASE_URL,
-  XENTE_ACCOUNT_ID: process.env.DEV_XENTE_ACCOUNT_ID,
-  XENTE_PASSWORD: process.env.DEV_XENTE_PASSWORD,
-  XENTE_USERNAME: process.env.DEV_XENTE_USERNAME,
-  XENTE_COLLECTIONS_PRODUCT_ITEM:
-    process.env.DEV_XENTE_COLLECTIONS_PRODUCT_ITEM,
-  XENTE_COLLECTIONS_PRODUCT_REFERENCE:
-    process.env.DEV_XENTE_COLLECTIONS_PRODUCT_REFERENCE,
-  XENTE_C0LLECTIONS_TYPE: process.env.DEV_XENTE_C0LLECTIONS_TYPE,
-  XENTE_PAYOUTS_PAYMENT_REFERENCE:
-    process.env.DEV_XENTE_PAYOUTS_PAYMENT_REFERENCE,
-  XENTE_PAYOUTS_TYPE: process.env.DEV_XENTE_PAYOUTS_TYPE,
-  XENTE_PAYOUTS_PAYMENT_PROVIDER:
-    process.env.DEV_XENTE_PAYOUTS_PAYMENT_PROVIDER,
-  XENTE_DATA_PAYMENT_PROVIDER: process.env.DEV_XENTE_DATA_PAYMENT_PROVIDER,
-  XENTE_DATA_PAYMENT_REFERENCE: process.env.DEV_XENTE_DATA_PAYMENT_REFERENCE,
-  XENTE_DATA_TYPE: process.env.DEV_XENTE_DATA_TYPE,
-  THINGS_MOBILE_RECHARGE_URL: process.env.DEV_THINGS_MOBILE_RECHARGE_URL,
-  THINGS_MOBILE_UPDATE_SIM_NAME_URL:
-    process.env.DEV_THINGS_MOBILE_UPDATE_SIM_NAME_URL,
-  THINGS_MOBILE_DEACTIVATE_URL: process.env.DEV_THINGS_MOBILE_DEACTIVATE_URL,
-  THINGS_MOBILE_ACTIVATE_URL: process.env.DEV_THINGS_MOBILE_ACTIVATE_URL,
-  THINGS_MOBILE_STATUS_URL: process.env.DEV_THINGS_MOBILE_STATUS_URL,
-  THINGS_MOBILE_BASE_URL: process.env.DEV_THINGS_MOBILE_BASE_URL,
-  THINGS_MOBILE_TOKEN: process.env.DEV_THINGS_MOBILE_TOKEN,
-  THINGS_MOBILE_USERNAME: process.env.DEV_THINGS_MOBILE_USERNAME,
-};
-const stageConfig = {
-  MONGO_URI: process.env.STAGE_MONGO_URI,
-  ENVIRONMENT: "STAGING ENVIRONMENT",
-  DB_NAME: process.env.STAGE_MONGO_DB,
-  REDIS_SERVER: process.env.STAGE_REDIS_SERVER,
-  REDIS_PORT: process.env.STAGE_REDIS_PORT,
-  XENTE_BASE_URL: process.env.STAGE_XENTE_BASE_URL,
-  XENTE_ACCOUNT_ID: process.env.STAGE_XENTE_ACCOUNT_ID,
-  XENTE_PASSWORD: process.env.STAGE_XENTE_PASSWORD,
-  XENTE_USERNAME: process.env.STAGE_XENTE_USERNAME,
-  XENTE_COLLECTIONS_PRODUCT_ITEM:
-    process.env.STAGE_XENTE_COLLECTIONS_PRODUCT_ITEM,
-  XENTE_COLLECTIONS_PRODUCT_REFERENCE:
-    process.env.STAGE_XENTE_COLLECTIONS_PRODUCT_REFERENCE,
-  XENTE_C0LLECTIONS_TYPE: process.env.STAGE_XENTE_C0LLECTIONS_TYPE,
-  XENTE_PAYOUTS_PAYMENT_REFERENCE:
-    process.env.STAGE_XENTE_PAYOUTS_PAYMENT_REFERENCE,
-  XENTE_PAYOUTS_TYPE: process.env.STAGE_XENTE_PAYOUTS_TYPE,
-  XENTE_PAYOUTS_PAYMENT_PROVIDER:
-    process.env.STAGE_XENTE_PAYOUTS_PAYMENT_PROVIDER,
-  XENTE_DATA_PAYMENT_PROVIDER: process.env.STAGE_XENTE_DATA_PAYMENT_PROVIDER,
-  XENTE_DATA_PAYMENT_REFERENCE: process.env.STAGE_XENTE_DATA_PAYMENT_REFERENCE,
-  XENTE_DATA_TYPE: process.env.STAGE_XENTE_DATA_TYPE,
-  THINGS_MOBILE_RECHARGE_URL: process.env.STAGE_THINGS_MOBILE_RECHARGE_URL,
-  THINGS_MOBILE_UPDATE_SIM_NAME_URL:
-    process.env.STAGE_THINGS_MOBILE_UPDATE_SIM_NAME_URL,
-  THINGS_MOBILE_DEACTIVATE_URL: process.env.STAGE_THINGS_MOBILE_DEACTIVATE_URL,
-  THINGS_MOBILE_ACTIVATE_URL: process.env.STAGE_THINGS_MOBILE_ACTIVATE_URL,
-  THINGS_MOBILE_STATUS_URL: process.env.STAGE_THINGS_MOBILE_STATUS_URL,
-  THINGS_MOBILE_BASE_URL: process.env.STAGE_THINGS_MOBILE_BASE_URL,
-  THINGS_MOBILE_TOKEN: process.env.STAGE_THINGS_MOBILE_TOKEN,
-  THINGS_MOBILE_USERNAME: process.env.STAGE_THINGS_MOBILE_USERNAME,
-};
-const prodConfig = {
-  MONGO_URI: process.env.PROD_MONGO_URI,
-  DB_NAME: process.env.PROD_MONGO_DB,
-  ENVIRONMENT: "PRODUCTION ENVIRONMENT",
-  REDIS_SERVER: process.env.PROD_REDIS_SERVER,
-  REDIS_PORT: process.env.PROD_REDIS_PORT,
-  XENTE_BASE_URL: process.env.PROD_XENTE_BASE_URL,
-  XENTE_ACCOUNT_ID: process.env.PROD_XENTE_ACCOUNT_ID,
-  XENTE_PASSWORD: process.env.PROD_XENTE_PASSWORD,
-  XENTE_USERNAME: process.env.PROD_XENTE_USERNAME,
-  XENTE_COLLECTIONS_PRODUCT_ITEM:
-    process.env.PROD_XENTE_COLLECTIONS_PRODUCT_ITEM,
-  XENTE_COLLECTIONS_PRODUCT_REFERENCE:
-    process.env.PROD_XENTE_COLLECTIONS_PRODUCT_REFERENCE,
-  XENTE_C0LLECTIONS_TYPE: process.env.PROD_XENTE_C0LLECTIONS_TYPE,
-  XENTE_PAYOUTS_PAYMENT_REFERENCE:
-    process.env.PROD_XENTE_PAYOUTS_PAYMENT_REFERENCE,
-  XENTE_PAYOUTS_TYPE: process.env.PROD_XENTE_PAYOUTS_TYPE,
-  XENTE_PAYOUTS_PAYMENT_PROVIDER:
-    process.env.PROD_XENTE_PAYOUTS_PAYMENT_PROVIDER,
-  XENTE_DATA_PAYMENT_PROVIDER: process.env.PROD_XENTE_DATA_PAYMENT_PROVIDER,
-  XENTE_DATA_PAYMENT_REFERENCE: process.env.PROD_XENTE_DATA_PAYMENT_REFERENCE,
-  XENTE_DATA_TYPE: process.env.PROD_XENTE_DATA_TYPE,
-  THINGS_MOBILE_RECHARGE_URL: process.env.PROD_THINGS_MOBILE_RECHARGE_URL,
-  THINGS_MOBILE_UPDATE_SIM_NAME_URL:
-    process.env.PROD_THINGS_MOBILE_UPDATE_SIM_NAME_URL,
-  THINGS_MOBILE_DEACTIVATE_URL: process.env.PROD_THINGS_MOBILE_DEACTIVATE_URL,
-  THINGS_MOBILE_ACTIVATE_URL: process.env.PROD_THINGS_MOBILE_ACTIVATE_URL,
-  THINGS_MOBILE_STATUS_URL: process.env.PROD_THINGS_MOBILE_STATUS_URL,
-  THINGS_MOBILE_BASE_URL: process.env.PROD_THINGS_MOBILE_BASE_URL,
-  THINGS_MOBILE_TOKEN: process.env.PROD_THINGS_MOBILE_TOKEN,
-  THINGS_MOBILE_USERNAME: process.env.PROD_THINGS_MOBILE_USERNAME,
-};
-const defaultConfig = {
-  NETWORK_INCLUSION_PROJECTION: {
-    description: 1,
-    name: 1,
-    _id: 1,
-  },
-  NETWORK_EXCLUSION_PROJECTION: (category) => {
-    const initialProjection = { nothing: 0 };
-    let projection = Object.assign({}, initialProjection);
-    if (category === "summary") {
-      projection = Object.assign({}, {});
-    }
-    return projection;
-  },
-  SESSION_SECRET: process.env.SESSION_SECRET,
-  XENTE_PAYMENT_PROVIDER: process.env.XENTE_PAYMENT_PROVIDER,
-  XENTE_METADATA: process.env.XENTE_METADATA,
-  XENTE_BATCH_ID: process.env.XENTE_BATCH_ID,
-  XENTE_REQUEST_ID: process.env.XENTE_REQUEST_ID,
-  XENTE_MEMO: process.env.XENTE_MEMO,
-  XENTE_CHANNEL_ID: process.env.XENTE_CHANNEL_ID,
-  XENTE_CUSTOMER_ID: process.env.XENTE_CUSTOMER_ID,
-  XENTE_CUSTOMER_PHONE: process.env.XENTE_CUSTOMER_PHONE,
-  XENTE_CUSTOMER_EMAIL: process.env.XENTE_CUSTOMER_EMAIL,
-  NETWORKS: process.env.NETWORKS
-    ? process.env.NETWORKS.split(",").filter((value) => value.trim() !== "")
-    : [],
-  SLACK_TOKEN: process.env.SLACK_TOKEN,
-  SLACK_CHANNEL: process.env.SLACK_CHANNEL,
-  SLACK_USERNAME: process.env.SLACK_USERNAME,
-  PORT: process.env.PORT || 3000,
-  JWT_SECRET: process.env.JWT_SECRET,
-};
+const environments = require("./environments");
+const global = require("./global");
 
 function envConfig(env) {
-  switch (env) {
-    case "development":
-      return devConfig;
-    case "staging":
-      return stageConfig;
-    default:
-      return prodConfig;
-  }
+  return { ...global, ...environments[env] };
 }
 
-module.exports = { ...defaultConfig, ...envConfig(process.env.NODE_ENV) };
+module.exports = envConfig(process.env.NODE_ENV);
