+++ src/incentives/config/environments/index.js
@@ -1,0 +1,11 @@
+const development = require("./development");
+const production = require("./production");
+const staging = require("./staging");
+
+const configurations = {
+  development,
+  production,
+  staging,
+};
+
+module.exports = configurations;
