+++ src/incentives/config/global/mappings.js
@@ -1,0 +1,4 @@
+const mongoose = require("mongoose");
+
+const mappings = {};
+module.exports = mappings;
