+++ src/incentives/config/global/numericals.js
@@ -1,0 +1,4 @@
+const mongoose = require("mongoose");
+
+const numericals = { PORT: process.env.PORT || 3000 };
+module.exports = numericals;
