+++ src/incentives/config/global/query-limits.js
@@ -1,0 +1,4 @@
+const mongoose = require("mongoose");
+
+const queryLimits = {};
+module.exports = queryLimits;
