+++ src/incentives/config/global/regex-patterns.js
@@ -1,0 +1,5 @@
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
+
+const regexPatterns = {};
+module.exports = regexPatterns;
