+++ src/incentives/config/global/slack.js
@@ -1,0 +1,8 @@
+const mongoose = require("mongoose");
+
+const credentials = {
+  SLACK_TOKEN: process.env.SLACK_TOKEN,
+  SLACK_CHANNEL: process.env.SLACK_CHANNEL,
+  SLACK_USERNAME: process.env.SLACK_USERNAME,
+};
+module.exports = credentials;
