+++ src/incentives/config/global/strings.js
@@ -1,0 +1,6 @@
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
+const isEmpty = require("is-empty");
+
+const strings = {};
+module.exports = strings;
