+++ src/incentives/config/global/urls.js
@@ -1,0 +1,5 @@
+const mongoose = require("mongoose");
+const ObjectId = mongoose.Types.ObjectId;
+
+const urls = {};
+module.exports = urls;
